comments and README

+++ src/main/java/com/nosto/api/dtos/ExchangeRateApiResponse.java
@@ -10,6 +10,17 @@ import lombok.Setter;
 import java.util.HashMap;
 import java.util.Map;
 
+/**
+ * Represents the answer provided by https://api.exchangeratesapi.io/latest?base={fromCcy}&symbols={toCcy}
+ * Example answer:
+ * {
+ *     "rates": {
+ *         "EUR": 0.222424876
+ *     },
+ *     "base": "BRL",
+ *     "date": "2019-08-20"
+ * }
+ */
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Data
 @NoArgsConstructor
@@ -19,14 +30,14 @@ public class ExchangeRateApiResponse {
     private String date;
 
     @Setter(AccessLevel.NONE)
-    private final Map<String, Float> rates = new HashMap<>();
+    private final Map<String, Double> rates = new HashMap<>();
 
     @JsonAnySetter
-    public void addCurrencyRate(String currencyCode, float rate) {
+    public void addCurrencyRate(String currencyCode, Double rate) {
         this.rates.put(currencyCode, rate);
     }
 
-    public float getCurrencyRate(String ccyCode) {
+    public Double getCurrencyRate(String ccyCode) {
         return rates.get(ccyCode);
     }
 }
